feat: export profile schema from another file to allow import on frontend
