+++ package.json
@@ -8,11 +8,10 @@
   "files": [
     "dist/types",
     "dist/src/**/*.dto.{js,d.ts,ts}",
-    "dist/src/**/constants.{js,d.ts,ts}",
     "dist/src/**/*.schema.{js,d.ts,ts}",
+    "dist/src/**/constants.{js,d.ts,ts}",
     "dist/src/**/exports.{js,d.ts,ts}",
     "dist/src/shared/dto-modification",
-    "dist/src/shared/errors/dotType.ts",
     "dist/src/i18n"
   ],
   "scripts": {
@@ -21,7 +20,7 @@
     "postinstall": "patch-package",
     "format": "prettier --write \"src/**/*.ts\" \"types/**/*.ts\"",
     "check-format": "prettier --check \"src/**/*.ts\"",
-    "check-exports": "attw --pack --profile node16 .",
+    "check-exports": "attw --pack .",
     "local-release": "changeset version && changeset publish",
     "commit-schema": "git add ./types/schema.d.ts swagger.yaml && git commit -m \"Update schema\"",
     "start": "npx prisma generate && npx nest start",
@@ -42,24 +41,37 @@
     "@anatine/zod-nestjs": "^2.0.9",
     "@anatine/zod-openapi": "^2.2.6",
     "i18n-js": "^4.4.3",
-    "libphonenumber-js": "^1.10.53",
     "patch-package": "^8.0.0",
     "zod": "^3.23.8"
   },
   "devDependencies": {
-    "@arethetypeswrong/cli": "^0.17.3",
-    "@changesets/cli": "^2.27.7",
-    "@nestjs/cli": "^10.0.0",
+    "reflect-metadata": "^0.2.0",
+    "string-comparison": "^1.3.0",
+    "json-to-pretty-yaml": "^1.2.2",
+    "jszip": "^3.10.1",
+    "multer": "^1.4.5-lts.1",
+    "openapi3-ts": "^4.4.0",
     "@nestjs/common": "^10.0.0",
     "@nestjs/config": "^3.2.3",
     "@nestjs/core": "^10.0.0",
     "@nestjs/jwt": "^10.2.0",
     "@nestjs/platform-express": "^10.0.0",
-    "@nestjs/schematics": "^10.0.0",
     "@nestjs/swagger": "^7.3.1",
+    "@prisma/client": "^6.3.1",
+    "bcrypt": "^5.1.1",
+    "body-parser": "^1.20.3",
+    "class-transformer": "^0.5.1",
+    "class-validator": "^0.14.1",
+    "country-state-city": "^3.2.1",
+    "date-fns": "^4.1.0",
+    "exceljs": "^4.4.0",
+    "fast-csv": "^5.0.2",
+    "@arethetypeswrong/cli": "^0.15.4",
+    "@changesets/cli": "^2.27.7",
+    "@nestjs/cli": "^10.0.0",
+    "@nestjs/schematics": "^10.0.0",
     "@nestjs/testing": "^10.0.0",
     "@openapitools/openapi-generator-cli": "^2.13.4",
-    "@prisma/client": "^6.3.1",
     "@types/bcrypt": "^5.0.2",
     "@types/express": "^4.17.17",
     "@types/jest": "^29.5.2",
@@ -70,33 +82,19 @@
     "@types/supertest": "^6.0.0",
     "@typescript-eslint/eslint-plugin": "^8.8.0",
     "@typescript-eslint/parser": "^8.8.0",
-    "bcrypt": "^5.1.1",
-    "body-parser": "^1.20.3",
-    "class-transformer": "^0.5.1",
-    "class-validator": "^0.14.1",
-    "country-state-city": "^3.2.1",
-    "date-fns": "^4.1.0",
+    "eslint": "^9.11.1",
     "eslint-config-prettier": "^9.0.0",
     "eslint-plugin-prettier": "^5.0.0",
     "eslint-plugin-unused-imports": "^4.1.4",
-    "eslint": "^9.11.1",
-    "exceljs": "^4.4.0",
-    "fast-csv": "^5.0.2",
     "globals": "^15.10.0",
     "husky": "^9.0.11",
     "jest": "^29.5.0",
-    "json-to-pretty-yaml": "^1.2.2",
-    "jszip": "^3.10.1",
     "lint-staged": "^15.2.5",
     "luxon": "^3.5.0",
-    "multer": "^1.4.5-lts.1",
     "openapi-typescript": "^7.0.0-rc.0",
-    "openapi3-ts": "^4.4.0",
     "prettier": "^3.0.0",
     "prisma": "^6.3.1",
-    "reflect-metadata": "^0.2.0",
     "source-map-support": "^0.5.21",
-    "string-comparison": "^1.3.0",
     "supertest": "^6.3.3",
     "ts-jest": "^29.1.0",
     "ts-loader": "^9.4.3",
