+++ src/otp/dto/send-otp.dto.ts
@@ -1,4 +1,4 @@
-import { profileSchema } from '@/profile/schema/profile.schema';
+import { profileSchema } from '@/schema/profile.schema';
 import { createZodDtoWithoutDate } from '@/shared/dto-modification/create-zod-dto-without-date';
 import z from 'zod';
