+++ src/otp/dto/verify-otp.dto.ts
@@ -1,6 +1,6 @@
 import { translate } from '@/i18n/translate';
 import { OTP_LENGTH } from '@/otp/constants';
-import { profileSchema } from '@/profile/schema/profile.schema';
+import { profileSchema } from '@/schema/profile.schema';
 import { createZodDtoWithoutDate } from '@/shared/dto-modification/create-zod-dto-without-date';
 import z from 'zod';
