+++ src/profile/dto/delete-profile.dto.ts
@@ -1,5 +1,7 @@
-import { deleteProfileResponseSchema } from '@/profile/schema/delete-profile.schema';
+import { profileSchema } from '@/schema/profile.schema';
 import { createZodDtoWithoutDate } from '@/shared/dto-modification/create-zod-dto-without-date';
+
+export const deleteProfileResponseSchema = profileSchema;
 
 export class DeleteProfileResponseDto extends createZodDtoWithoutDate(
   deleteProfileResponseSchema,
