+++ src/profile/dto/find-by-phone-number.dto.ts
@@ -1,5 +1,7 @@
-import { findByPhoneNumberResponseSchema } from '@/profile/schema/find-by-phone-number.schema';
+import { profileSchema } from '@/schema/profile.schema';
 import { createZodDtoWithoutDate } from '@/shared/dto-modification/create-zod-dto-without-date';
+
+export const findByPhoneNumberResponseSchema = profileSchema;
 
 export class FindByPhoneNumberResponseDto extends createZodDtoWithoutDate(
   findByPhoneNumberResponseSchema,
