+++ src/profile/dto/find-by-tag-groups-profile.dto.ts
@@ -1,5 +1,17 @@
-import { findByTagGroupsProfileResponseSchema } from '@/profile/schema/find-by-tag-groups-profile.schema';
+import { profileSchema } from '@/schema/profile.schema';
 import { createZodDtoWithoutDate } from '@/shared/dto-modification/create-zod-dto-without-date';
+import { tagSchema } from '@/tag/dto/tag.dto';
+import z from 'zod';
+
+export const findByTagGroupsProfileResponseSchema = z.object({
+  profiles: z.array(
+    profileSchema.merge(
+      z.object({
+        tags: z.array(tagSchema),
+      }),
+    ),
+  ),
+});
 
 export class FindByTagGroupsProfileResponseDto extends createZodDtoWithoutDate(
   findByTagGroupsProfileResponseSchema,
