+++ src/schema/exports.ts
@@ -1,0 +1,1 @@
+export * from './profile.schema';
