+++ swagger.yaml
@@ -6239,14 +6239,6 @@ components:
               type:
                 - "string"
                 - "null"
-            password:
-              type:
-                - "string"
-                - "null"
-            username:
-              type:
-                - "string"
-                - "null"
             comments:
               type:
                 - "array"
@@ -6340,8 +6332,6 @@ components:
             - "phoneNumber"
             - "profilePictureUrl"
             - "secondaryPhoneNumber"
-            - "password"
-            - "username"
         checkForSimilarity:
           type:
             - "boolean"
@@ -6880,14 +6870,6 @@ components:
             - "string"
             - "null"
           format: "date-time"
-        username:
-          type:
-            - "string"
-            - "null"
-        password:
-          type:
-            - "string"
-            - "null"
         residence:
           type:
             - "object"
